+++ CHANGELOG.md
@@ -7,6 +7,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 An entry should be added to the _[Unreleased]_ section when committing to main. When a release is made we move its
 contents to a new release section (e.g. _[0.x]_).
 
+## [Unreleased]
+
+### Added
+
+### Fixed
+
+### Changed
+
+### Removed
+
+## [1.4.1] 2023-12-22
+
+### Changed
+
+- Normalize polynomials to a stricter standard form (fractional coefficients in front of variables) (PLUT-943)
+- gmAction for adding a fraction and a non-fraction
+
 ## [1.4] 2023-12-21
 
 ### Added